+++ frontend/halyk_ambassador_client/lib/features/auth/domain/entities/profile_entities.dart
@@ -84,7 +84,7 @@ class Address extends Equatable {
   final String postcode;
   final double latitude;
   final double longitude;
-  final int confidence;
+  final double confidence;
 
   const Address({
     required this.found,
